Separate checkout-only base_cli tests
