+++ tests/test_base_cli_public_command_lifecycle.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 
 
-REPO_ROOT = Path(__file__).resolve().parents[4]
+REPO_ROOT = Path(__file__).resolve().parents[1]
 PUBLIC_COMMAND_ROOT = REPO_ROOT / "cli" / "python"
