+++ tests/test_base_cli_python_standards.py
@@ -4,7 +4,7 @@ import ast
 from pathlib import Path
 
 
-REPO_ROOT = Path(__file__).resolve().parents[4]
+REPO_ROOT = Path(__file__).resolve().parents[1]
 STANDARDS_DOC = REPO_ROOT / "STANDARDS.md"
 PYTHON_ROOTS = (
     REPO_ROOT / "cli" / "python",
